pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
    "mysqlclient>=2.2.8",
]
